bundles: Rename commons.js to common.js.
